create intro file
